requirements.txt
discord.py
yt-dlp
gtts
flask
requests
